+++ server/static/index.html
@@ -111,7 +111,7 @@
                 for (let roleID in roles) {
                     let role = roles[roleID];
                     let roleDefinition = roleDefinitions[role.name];
-                    console.debug(roleID, role, roleDefinition);
+                    // console.debug(roleID, role, roleDefinition);
 
                     let roleButton = document.createElement("input");
                     roleButton.id = roleID;
